Use Float instead of Double

+++ src/main/kotlin/fi/evident/fab/Types.kt
@@ -1,5 +1,5 @@
 package fi.evident.fab
 
-typealias Hz = Double
-typealias dB = Double
+typealias Hz = Float
+typealias dB = Float
